refactor: upload on activists page
