+++ frontend-v2/src/app/(authed)/activists/activists-page.tsx
@@ -1,9 +1,8 @@
 'use client'
 
-import { useEffect, useMemo, useRef, useState, useCallback } from 'react'
+import { useMemo, useState } from 'react'
 import { useInfiniteQuery, useQuery } from '@tanstack/react-query'
 import { useQueryState, parseAsInteger } from 'nuqs'
-import toast from 'react-hot-toast'
 import {
   apiClient,
   API_PATH,
@@ -24,6 +23,7 @@ import { buildQueryOptions } from './filter-api-query'
 import type { ActivistsQueryState, SortColumn } from './query-state'
 import { DEFAULT_SORT } from './query-state'
 import { useActivistQueryState } from './use-activist-query-state'
+import { ExportButton } from './export-button'
 
 interface ActivistsPageProps {
   debugInitialServerQueryState?: ActivistsQueryState
@@ -69,18 +69,6 @@ export default function ActivistsPage({
     columns: selectedColumns,
     sort,
   })
-
-  const [isExporting, setIsExporting] = useState(false)
-  const exportAbortControllerRef = useRef<AbortController | null>(null)
-
-  useEffect(() => {
-    const controller = new AbortController()
-    exportAbortControllerRef.current = controller
-    return () => {
-      controller.abort()
-      exportAbortControllerRef.current = null
-    }
-  }, [])
 
   const isExplicitSort = sort.length > 0
   const effectiveSort = isExplicitSort ? sort : DEFAULT_SORT
@@ -165,72 +153,6 @@ export default function ActivistsPage({
       apiClient.countActivists(countQueryOptions, signal),
   })
 
-  // The spoke export uses the current filters but a server-selected column
-  // set, so we send an empty columns array. The server hard-codes the spoke
-  // columns and rejects a non-empty list.
-  const spokeQueryOptions = useMemo<QueryActivistOptions>(
-    () => ({
-      ...queryOptions,
-      shape: { ...queryOptions.shape, columns: [] },
-    }),
-    [queryOptions],
-  )
-
-  const runExport = useCallback(
-    async (
-      fetchBlob: (signal: AbortSignal) => Promise<Blob>,
-      filenamePrefix: string,
-    ) => {
-      if (isExporting) return
-      const controller = exportAbortControllerRef.current
-      if (!controller) return
-      const { signal } = controller
-      setIsExporting(true)
-      let url: string | undefined
-      try {
-        const blob = await fetchBlob(signal)
-        if (signal.aborted) return
-        url = URL.createObjectURL(blob)
-        const a = document.createElement('a')
-        a.href = url
-        a.download = `${filenamePrefix}-${new Date().toISOString().slice(0, 10)}.csv`
-        document.body.appendChild(a)
-        a.click()
-        a.remove()
-      } catch (err) {
-        if (signal.aborted) return
-        console.error('Failed to export activists CSV', err)
-        toast.error(
-          err instanceof Error && err.message
-            ? `Failed to export activists: ${err.message}`
-            : 'Failed to export activists. Please try again.',
-        )
-      } finally {
-        if (url) URL.revokeObjectURL(url)
-        if (!signal.aborted) setIsExporting(false)
-      }
-    },
-    [isExporting],
-  )
-
-  const handleExport = useCallback(
-    () =>
-      runExport(
-        (signal) => apiClient.exportActivistsCsv(queryOptions, signal),
-        'activists',
-      ),
-    [runExport, queryOptions],
-  )
-  const handleExportSpoke = useCallback(
-    () =>
-      runExport(
-        (signal) =>
-          apiClient.exportActivistsSpokeCsv(spokeQueryOptions, signal),
-        'activists-spoke',
-      ),
-    [runExport, spokeQueryOptions],
-  )
-
   const activists: ActivistJSON[] = useMemo(
     () => data?.pages.flatMap((page) => page.activists) ?? [],
     [data],
@@ -269,9 +191,7 @@ export default function ActivistsPage({
           isAdmin={isAdmin}
           isDirty={isDirty}
           onReset={resetAll}
-          onExport={handleExport}
-          onExportSpoke={handleExportSpoke}
-          isExporting={isExporting}
+          exportButton={<ExportButton queryOptions={queryOptions} />}
         >
           <ColumnSelector
             visibleColumns={selectedColumns}
