+++ frontend-v2/src/app/(authed)/activists/filters/activist-filters.tsx
@@ -7,7 +7,7 @@ import {
   PopoverContent,
   PopoverTrigger,
 } from '@/components/ui/popover'
-import { ChevronDown, Download, Plus, RotateCcw, X } from 'lucide-react'
+import { Plus, RotateCcw, X } from 'lucide-react'
 import { ActivistLevelFilter } from './activist-level-filter'
 import { LastEventFilter } from './last-event-filter'
 import { DateRangeFilter } from './date-range-filter'
@@ -24,9 +24,7 @@ interface ActivistFiltersProps {
   isAdmin: boolean
   isDirty: boolean
   onReset: () => void
-  onExport?: () => void
-  onExportSpoke?: () => void
-  isExporting?: boolean
+  exportButton?: React.ReactNode
   children?: React.ReactNode
 }
 
@@ -61,9 +59,7 @@ export function ActivistFilters({
   isAdmin,
   isDirty,
   onReset,
-  onExport,
-  onExportSpoke,
-  isExporting,
+  exportButton,
   children,
 }: ActivistFiltersProps) {
   // Tracks optional filters added from the menu that may not yet have values.
@@ -71,7 +67,6 @@ export function ActivistFilters({
     new Set(),
   )
   const [isAddFilterOpen, setIsAddFilterOpen] = useState(false)
-  const [isExportMenuOpen, setIsExportMenuOpen] = useState(false)
   const canReset = isDirty || visibleFilters.size > 0
 
   // Prune visibleFilters entries that have no corresponding filter value when
@@ -143,48 +138,7 @@ export function ActivistFilters({
 
       {/* Chip bar */}
       <div className="flex flex-wrap items-center gap-2">
-        {onExport && (
-          <Popover open={isExportMenuOpen} onOpenChange={setIsExportMenuOpen}>
-            <PopoverTrigger asChild>
-              <Button
-                variant="ghost"
-                size="sm"
-                className="h-12 gap-1"
-                disabled={isExporting}
-              >
-                <Download className="h-4 w-4" />
-                {isExporting ? 'Exporting…' : 'Export'}
-                <ChevronDown className="h-4 w-4" />
-              </Button>
-            </PopoverTrigger>
-            <PopoverContent className="w-44 p-2" align="start">
-              <div className="flex flex-col">
-                <button
-                  type="button"
-                  className="flex w-full items-center rounded px-2 py-1.5 text-sm hover:bg-muted transition-colors text-left"
-                  onClick={() => {
-                    setIsExportMenuOpen(false)
-                    onExport()
-                  }}
-                >
-                  Current columns
-                </button>
-                {onExportSpoke && (
-                  <button
-                    type="button"
-                    className="flex w-full items-center rounded px-2 py-1.5 text-sm hover:bg-muted transition-colors text-left"
-                    onClick={() => {
-                      setIsExportMenuOpen(false)
-                      onExportSpoke()
-                    }}
-                  >
-                    Spoke columns
-                  </button>
-                )}
-              </div>
-            </PopoverContent>
-          </Popover>
-        )}
+        {exportButton}
         <Button
           variant="ghost"
           size="sm"
